reversed order of id and item type

+++ src/main/java/com/zipcodewilmington/heroku_demo/City.java
@@ -8,7 +8,7 @@ public class City {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String name;
     private int population;
@@ -16,17 +16,17 @@ public class City {
     public City() {
     }
 
-    public City(Integer id, String name, int population) {
+    public City(Long id, String name, int population) {
         this.id = id;
         this.name = name;
         this.population = population;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
